Create discord.js and sleeper.js  for isolate code and another moment can be use this code. Resolve problem in code responsable send jobs in discord
