+++ backend/src/utils/sleeper.js
@@ -1,0 +1,10 @@
+
+const sleep = (seconds) => {
+    return new Promise((resolve, reject) => {
+        setTimeout(() => resolve(), seconds * 1000);
+    });
+};
+
+module.exports = {
+    sleep
+}
